requirements.txt
pocketflow>=0.0.1
pyyaml>=6.0
requests>=2.28.0
gitpython>=3.1.0
python-dotenv>=1.0.0
pathspec>=0.11.0
litellm>=1.37.6
